drivers/ob-bigquery/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ob-bigquery"
version = "0.1.0"
description = "OrionBelt Semantic Layer driver for BigQuery (DB-API 2.0)"
license = { text = "Apache-2.0" }
authors = [{ name = "Ralfo Becher", email = "info@orionbelt.ai" }]
keywords = ["bigquery", "semantic-layer", "obml", "db-api", "orionbelt"]
requires-python = ">=3.12"
dependencies = [
    "ob-driver-core>=0.1",
    "google-cloud-bigquery[pandas]>=3.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.10",
    "ruff>=0.4",
    "respx>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ob_bigquery"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ANN"]

[tool.mypy]
python_version = "3.12"
strict = true
files = ["src/ob_bigquery"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = ["integration: requires live DB + OB API"]
addopts = "-m 'not integration'"
